unify setup scripts + add pipenv monorepo (#273)
Adding `yamllint` in the following PR, and turned out that mono-repo support in Python is ... dissappointing.
Struggled for a while to find a good multi-package manager that I can just import and use like how we use Cargo and NPM workspaces, then gave up.
This PR changes our setup scripts to loop over individual packages and install them in a loop, which is a simple solution for now.

Additionally, this fixes an issue where multiple crates are building at the same time, and each trying to invoke `setup.sh` if `prettier` is not found.
The issue is compounded as `npm install` is not parallel-safe. So I chose to error out instead.

@AntonyBlakey since you mentioned that breaks devcontainers when they are rebuilt,
I also added a `postAttachCommand` so that a new container should have a fresh installation on all packages. It will also warmup `rust-src` and `rust-analyzer`, since you mentioned they cause troubles in fresh dev containers. Please let me know if you have concerns on this new workflow.

Co-authored-by: Omar Tawfik <OmarTawfik@users.noreply.github.com>