use `number` and getters in npm public API
Follow up on #458

Turns out that using `#[napi(ts_return_type = "number")]` only affects the return type in the generated `index.d.ts` file, but not the actual return type at runtime.

Now we are using `usize`, which maps to JavaScript's `BigInt` by default, which is not ideal to work with at all. Using `u32` instead which maps to JavaScript's `number` type.

Also, adds a few missing `getter` attributes, and adds a test to check the new types.