Drive-by tweaks and docs (#561)
Not touching the grammar definition-related code since that's about to
be changed by
https://github.com/AntonyBlakey/slang/tree/AntonyBlakey/to_rust_types.

Somewhat related to https://github.com/NomicFoundation/slang/issues/234,
I noticed that we have public extension traits that seem to only work
with the `LanguageDefinition` type. Maybe we should change these
one-function traits to free functions that accept the language
definition as the argument, instead?

Also not feeling strongly about
https://github.com/NomicFoundation/slang/commit/84912e8420b12c39074f57e96f576ee3e194d0a5,
thought it's more of a nice-to-have.