Attempt error recovery in the TerminatedBy parsers (#567)
Based on #564

EDIT:
As asked, this was split into supporting PRs (#579, #580, merged now),
doesn't use the Error CST node anymore and only contains terminated-by
recovery.

To support error side-channel in a backtracking scenario,
`Vec<ParseError>` was added to the `Stream` which serves more as a parse
context now. To backtrack, we first record the position of the stream
and how many errors are there; then, we reset the errors and the
position to the first recorded `Marker` struct (basically copying what
`chumsky` does).

This wasn't changed everywhere as `Stream::set_position` in the lexer
will never emit errors (only parsers do), nor will the optional trivia
parsers. I'm happy to polish the lexer interface a bit to accommodate
for this change later, if that's okay. First, I wanted to make sure the
approach is fine and accepted, until we proceed with the DelimitedBy
error recovery.