Improve error recovery in delimited groups when can't parse anything
TerminatedBy can be a bit too greedy, i.e. we recovered some unexpected
item definition just because... we recovered at a terminator. However,
the delimited group is a reasonable boundary since it's simple (2-3 kinds
of delimiters) and the opening delimiter must always match the closing
one, so it produces more reliable and self-contained recovered phrases,
so attempting to recover from a null parse in a delimited group makes
more sense.