Attempt error recovery from partial matches in separated lists (#595)
Based on #593.

Implements a dedicated `SeparatedHelper` (no state, can change to a free
function) that implements recovering from partial matches in separated
lists.

There are 3 ways we can recover:
1. a separatee matches but unexpected tokens follow before a separator
2. a separatee is partially matched
3. a separatee can't be parsed

Ideally we should recover in all three cases, however because the the
trailing separator is not mandatory (nor legal), cases 1. and 3. are
tricky. In case 2 The separated-by lists are not always in a delimited
group (e.g. `IdentifierPath`) and so there's no sure way to bound the
recovery, as the unexpected tokens (i.e. not a separator) can be already
part of the outer parse.

For instance, `0.0.0 || 0.0.0` (case 1) or `0.0. || 0.0.0` (case 3)
might attempt to recover by skipping the ` || 0` part. Until we
constrain the error recovery with a la FOLLOW for the recovering parser,
we can't reliably recover errors for now in these cases.