correct keyword versioning (#622)
- Split every `KeywordItem` into multiple `KeywordDefinition`s, each
having their own versioning. This is to account for things like
`fixed8x8` and `fixed256x80` being reserved/unreserved in different
versions.
- Duplicate keywords between Solidity and Yul, since the same keyword
can be reserved in different versions between the two.
- Now all keywords have correct versioning.
- Added a script that will fetch `solc` binaries and test all variants
of keywords (10777) against the actual compiler output. It has a few
hacks around parsing/skipping errors, so it is not stable enough (yet)
to run in CI/unguided, but it definitely helped running locally and
observing the output.