Inline the (Yul)Statement parser in v0/v1 (#658)
Part of #652

Seems convenient for the AST/CST (to not need to unwrap another layer)
but we lose the ability to test it as a named parser and the generated
parser code gets a bit bloated, since we now inline the choice.

Maybe it'd be better if we somehow allow to "outline" select enums in
v2, like this one? @OmarTawfik