chore: Specify `Optional(Required(List))` in v2 where we do in v1
Let's unify these definitions to be in line with what's in v0/v1 as to
not change the definition and the CST shape once we switch to using v2
as our source of truth for the parser.

This is independent whether we will want that *eventually* or not. For
now, we should keep things consistent during the migration period. After
that, we can make a decision whether we want to keep these empty nodes
in the CST or not in the absence of any separated items.