Move pragma-specific keywords under dedicated topic (#661)
Split from #650 

We talked about this recently at our stand-up.

These are never reserved in the default context and are only used in the
(lexical) context of pragmas, so let's reflect that in the spec.

Not updating v1 in this context since #650 is about to obsolete that and
v0 does not have lexical contexts that might be impacted.