Remove the CST Visitor API (#665)
We discussed internally that the Visitor API for (dynamically typed) CST
is a lot less useful now that we have the `Iterator` interface
implemented in Rust (and hopefully
[`Iterator`](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Guide/Iterators_and_Generators#iterators)
protocol in TS in the future), so let's remove the CST Visitor API
altogether.

The Visitor
[pattern](https://rust-unofficial.github.io/patterns/patterns/behavioural/visitor.html)
will be a lot useful once we have typed, heterogenous data such as the
new AST (#634), so let's wait until then with designing and exposing
such interface.