Backport `{Uint,Int,Fixed,Ufixed}Keyword` from DSL v2 to v0/v1 (#662)
Part of #652 

Helps with #650 

These make more sense, given that these are the names used in the
official grammar (fixed-bytes, signed-integer-type,
unsigned-integer-type) and "XXXKeyword" name implies a single atom and
not a more complex rule (e.g. looking at the `fixed` definition...).

I think we should revert the accidental name change in v2 and later
decide if we want to change these.

EDIT: This now does the opposite and backports the name change to v0/v1.