unify DSL model `Spanned<>` type tree
Started working on migrating DSL v0 to v2, and wanted to add common helpers/utilities to the model.
But faced the limitation is that the existing `codegen_language_internal_macros` requires defining
all model types in a single module, to allow generating a parallel `Spanned<>` type tree.
We no longer need to do that, since we are not parsing the DSL using `serde`.

This PR refactors `codegen_language_internal_macros` into more Rust-idiomatic `#[derive()` macros,
and removes the above restriction, allowing us to split model types into multiple modules as-needed.