Appease most of the Clippy lints (#673)
Part of #155 

These should all be more or less obvious, maybe with the exception of:
- ~`comparison_chain` - in this case it might be more noise than it's
worth, I'm fine with silencing it here, instead~
  EDIT: Actually, let's drop it for now
- `single_char_pattern` - in our case we do less `"`-escaping, so that's
good enough for me (but I know it might be considered controversial)
- `clippy::upper_case_acronyms` - this helps follow the Rust naming
convention (`CLI` -> `Cli`, which I'm strongly in favour; it's not a
SCREAMING_CASE_CONSTANT)

Some select lints are explicitly allowed on a case-by-case basis here
but in general are worthwhile to be linted against, e.g.
`enum_variant_names` or `manual_range_contains`.

The only ones left now are:
- `comparison_chain` - in this case it might be more noise than it's
worth, I'm fine with silencing it here, instead
- `needless_return`, which we will have *many* instances of and I'd like
to have a separate PR for that (but I'm *strongly* in favour, as I find
it to be one of the signature aspects of Rust)
- `should_implement_trait` - needs to be silenced in one case for
`#[napi] fn clone`, but `ParserContext::next` might or might not
implement the `Iterator`, I'm not sure yet, so I'd like to defer it for
now.