split DSL model types into multiple modules (#680)
Started working on migrating DSL v0 to v2, and wanted to add common
helpers/utilities to the model.
But faced the limitation is that the existing
`codegen_language_internal_macros` requires defining
all model types in a single module, to allow generating a parallel
`Spanned<>` type tree.
We no longer need to do that, since we are not parsing the DSL using
`serde`.

This PR refactors `codegen_language_internal_macros` to allow splitting
the model types,
at the expense of maintaining an allow-list of type names in
`internal_macros/src/derive/spanned.rs`.
I doubt we will be editing this list frequently, so it I think it should
be fine.