doc: Add an introduction and a design overview to the user guide (#674)
In preparation for #334 (and other how-tos).

This adds a bit more introduction and context to the user guide that
should be independent on the actual API used (Rust/TS) and also provides
stubs for the How-Tos on the parsing/navigating the syntax tree.

When I tried to write these specific how-tos, I thought that we should
be able to define, at least from the higher level, what the CST is, how
the process looks like and add some more context (e.g. multiple versions
and why we need to create a specific version for the language, first).
Also this helps establish a bit more structure to the guide that should
help with writing more documentation/guides as we go.

Ironically, http://shader-slang.com/slang/user-guide/index.html helped
me structure this a bit as a guide that the user is welcomed to and what
they might benefit from (intro, context) before diving deeper into the
guides/context.