Appease the `single_use_lifetimes` lint
Unfortunately, we can't turn it on because it has false positives that
suggest a code that fails to compile, i.e.
```rust
enum Errors<'err> {
    #[error("An item with the name '{0}' already exists.")]
    ExistingItem(&'err Identifier),
    #[error("A variant referencing '{0}' already exists.")]
    ExistingVariant(&'err Identifier),
    #[error("An expression with the name '{0}' already exists.")]
    ExistingExpression(&'err Identifier),
}
```

suggests to remove the `'err` but it's impossible to define the enum
using the elided lifetime, i.e. `enum Errors<'_>`.

See https://github.com/rust-lang/rust/issues/117965 for the existing bug report.