Warn against select few more `rustc` lints
The first one is enabled in the `rustc` itself and vendored-in tools:
`unused_lifetimes`. It makes sense to warn against it by default and
pepper allows or change the code such that it's used, e.g. with
PhantomData.

The second one will be already uplifted to `dead_code` (which is
warn-by-default) but since we're warning against some lints, it can't
hurt.