Add final Clippy lints (#704)
Final Clippy part of #155

## Lints

From the additional lint list mentioned by @OmarTawfik:
- Cargo - added everything, except the duplicate packages lint, as this
is not fixable and we should strive to periodically do that ourselves.
In any case, it's not something we should lint against and deny in CI
- Restriction:
- `clippy::same_name_method` - sounds useful but std traits are covered
already by `should_implement_trait` and this fires on wrapper
functions/traits; only 4 occurences in our codebase are intentional
wrappers
- `clippy::allow_attributes_without_reason` - requires a nightly rustc
feature
- `clippy::self_named_module_files` - forcing `mod.rs` onto users seems
backwards, as that's one of the key issues addressed for Rust 2018 in
terms of module/file path ergonomics, see
https://dev-doc.rust-lang.org/stable/edition-guide/rust-2018/path-changes.html#no-more-modrs
for more context
- `clippy::format_push_string` - alternative proposed is fallible,
whereas the "offending" code is infallible, not enough value IMO
- `clippy::str_to_string` - will be better fixed with `implicit_clone`
once we bump Rust to 1.7{3,4} with a new Clippy that has less false
positives
- `clippy::wildcard_enum_match_arm` - we already have
`match_wildcard_for_single_variants` + can have more false positives
- Nursery:
  - `clippy::use_self` - unstable, false positives
  - `clippy::collection_is_never_read` ✅ added
  - `clippy::equatable_if_let` ✅ added
  - `clippy::useless_let_if_seq` ✅ added
  
  
## Panicking
The remaining items from #155 are about panicking in N-API-exposed
crate.

Unfortunately, there is no way to guarantee or more broadly verify that
certain code paths don't panic, as `rustc` only compiles a single crate
at a time and panicking is not encoded in the type system. We could
brainstorm how we can make this more resilient on the Rust side but that
requires more data (maybe we could plug in Sentry for the Rust side of
VSCode extension?) and a more careful thinking rather than trying to
tackle this as part of the "Lint Rust code" item.