Add How-to guides on parsing Solidity for CLI/Rust/NPM (#716)
Closes #334 

This adds a more hands-on intro to the parsing API and adds two examples
on how to use it:
- how to reconstruct the source: it's a kind of a "hello world" since
you only visit and print the contents and it helps reinforce that the
CST is, in fact, complete (wrt trivia)
- how to list contracts and their names - it's simple, real worldsy and
uses two different node kinds in a simple relationship.

This is done for Rust CLI, Rust library and the NPM package used as a
library.

With this, we could probably get rid of the content from the respective
index pages for Rust/NPM package and use this as the introductory first
subpage as it overlaps with it now a bit.

Let me know if this is the direction we want to go in; didn't want to
add more involved examples until I was sure that's what we want.