Remove `EndOfFileTrivia` and always collect trivia after top-level parsing (#733)
Checks boxes in #640 #638

This adapts `ParserFunction` to always attempt parsing the leading
trivia after we had a succesful match. This has a nice side benefit of
not overwriting tokens that would allow for more progress when
collecting the final trivia (unlike previous [AttemptedParseRule,
EndOfFileTrivia] sequence).

I'm not so happy with the resulting tree shape, however; IMO we should
either:

1. return a "tree" that has both parents - the first one being the
expected parsed rule node but also the second parent being the "leading
trivia"
2. or at least introduce a notion of a virtual EndOfFile token that this
final end-of-file trivia can belong to.

Otherwise, a lone, final LeadingTrivia child belonging to the rule node
seems misplaced and it's not intuitive for me to find the remaining
trivia in that place. I'm open for other suggestions!