Rename `solidity_npm_crate` to `slang_solidity_node_addon`
To make it clearer that they're essentially sharing the source code and
that the sole reason for the "npm_crate" is to build a Node addon
cdylib.

I was always confused how `slang_solidity` is not `solidity_cargo_crate`
and that `solidity_npm_crate` is essentially `slang_solidity` but
slightly repackaged from the same source. This should make it clearer.

When/If we start exporting to Python, this should allow for a
clearer `slang_solidity_python_extension` or a more general
`slang_solidity_${LANG}_cdylib` name.