Add Tree Query Language and Engine (#753)
This pull request adds the CST Tree Query Language.

It is a long way from being optimised, and in particular it should not
need to allocate memory whilst backtracking, which it currently does.

It is built with understandability as the top priority.

This also includes some other changes that are not quite drive-by e.g. I
rationalised the runtime crate structure and file naming to make more
sense given I was adding a structural sub-module.