fix: Handle Skipped unwinding in Pratt parsing in RepetitionHelper
I didn't think we would encounter this use case; for now, we flatten
accumulated Pratt-parsed nodes and continue unwinding.

This might not be the optimal strategy, i.e. in the regression test it
would be better to ignore the Skipped and rely on "skip unrecognized
input" logic to handle it but we should be more consistent here and
continue unwinding if we encounter the unwinding SkippedUntil ourselves.