Don't panic when recovery-unwinding during Pratt-parsing (#794)
Fixes #788

Thanks @OmarTawfik for finding it!

I didn't think we would encounter this use case; for now, we flatten
accumulated Pratt-parsed nodes and continue unwinding.

This might not be the optimal strategy, i.e. in the regression test it
would be better to ignore the Skipped and rely on "skip unrecognized
input" logic to handle it (this would give us correct
`MemberAccessExpression` rather than `TypeName`) but we should be more
consistent here and continue unwinding if we encounter the unwinding
SkippedUntil ourselves; we can tweak recovery later.