Streamline cdylib setup for Node.js addon (#745)
Follow up to #735 

I scratched the NAPI build itch, as I never fully understood the root
issue. As I did, in the meantime I tried to better document and show
what the core issue really is
(https://github.com/rust-lang/cargo/issues/8628), so that future readers
do not have to get down that rabbit hole themselves.

Unfortunately, it's not entirely possible to eliminate the workaround
without patching napi-rs to support a different workaround (build cdylib
as example; see the link above). However, I tried to eliminate some of
the redundant setup not to confuse what's actually needed; for example,
we never build codegen_parser_runtime into `cdylib` but only use the
source inside and similarly the `solidity_npm_crate` (renamed to
`slang_solidity_node_addon` in
https://github.com/NomicFoundation/slang/commit/8853229c8f8db1966d4a4926b5ea3fd79afc21f7)
can only be ever compiled as `cdylib` and is not intended to be used as
a regular Rust `rlib`.

See the comments added in the diff and in the commits for more context.