fix: Ensure that Node.js native addon targets GLIBC 2.28 on `x86_64-unknown-linux-gnu` (#909)
Required for
https://github.com/NomicFoundation/hardhat-vscode/issues/546 #639

Without this, we host-compiled our native addon with the GLIBC of the
runner (`ubuntu-22.04`), which is 2.33. That's too high and will cause
linking issues on older, stable distributions such as Debian 11 (2.31),
Debian 10 (2.28), Ubuntu 20.04 (2.31) etc.

The Linux requirement for VS Code atm is 2.28
(https://code.visualstudio.com/docs/supporting/requirements#_additional-linux-requirements)
and was bumped in [1.86](https://code.visualstudio.com/updates/v1_86).
Prior to that, the required version was 2.17 and we might consider
targeting it instead, until we bump the required VS Code engine for the
shipped extension (cc @kanej).

Here is the latest runs that check that the `infra publish npm
--dry-run` executes as expected and passes the relevant checks on our
CI:
https://github.com/Xanewok/slang/actions/runs/8472800732/job/23215682564.
The built artifacts are uploaded as part of the pipeline so they can be
additionally downloaded and inspected manually for the GLIBC symbols.