support string literals in version pragmas (#839)
- the new grammar makes sure the structure of string literals matches
the parent productions, so that it can be parsed using the same parser
for later analysis.
- also fixed a minor grammar bug where `||` was a precedence binary
expression, instead of a separator of comparator sets (repeated list of
base expressions).
- removed the internal `extract_version_pragmas` API for now, since it
is not used. Later, we will introduce the same functionality as a public
API, based on the Rust AST when is ready.